Veículo DAO na vdd estava fazendo oque um DTO deveria fazer
